refactor: remove report service barrel reexports
